fix(122): fix CardPricing component and refactor pricing data structure
