+++ apps/xi.land/components/cardPricing/dataForPricing.ts
