+++ apps/xi.land/components/cardPricing/index.ts
@@ -1,0 +1,3 @@
+export { CardPricing } from './CardPricing';
+export type { CardPricingPropsT } from './dataForPricing';
+export { plansPricing, availableFeatures, pricingFaq, comparisonRows } from './dataForPricing';
